GmailPluginRoot/automation/requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0

